fixed compilation error on gcc 9

+++ src/gmail/GmailCache.cpp
@@ -3303,7 +3303,7 @@ void mail_cache::GThreadsStorage::bindSQL(QSqlQuery* q, CACHE_LIST<ThreadData>& 
         history_id << t->m_history_id;
         messages_count << t->m_messages_count;
         snippet << t->m_snippet;
-        labelsBitMap << t->labelsBitMap();
+        labelsBitMap << static_cast<quint64>(t->labelsBitMap());
         internalDate << t->internalDate();
         id << t->m_id;
 
@@ -3325,7 +3325,7 @@ bool mail_cache::GThreadsStorage::execOutOfBatchSQL(QSqlQuery* q, mail_cache::Th
     q->addBindValue(t->m_history_id);
     q->addBindValue(t->m_messages_count);
     q->addBindValue(t->m_snippet);
-    q->addBindValue(t->labelsBitMap());
+    q->addBindValue(static_cast<quint64>(t->labelsBitMap()));
     q->addBindValue(t->internalDate());
     q->addBindValue(t->m_id);
     return q->exec();
